Split ButtplugController into KK specific and reusable components
